Several things
- removed unnecessary else (regfree) clause in all functions and put
regfree outside the "if" clause

-  put memory allocation errors handling if's wherever I could find. Now
allocation errors are handled like it says in the exercise paper: print
a message and quit the program.

-fixed: when defining polynomials like p1 = 5x+4, it would catch
compound summation instead of define. (Same thing for subtraction).

-Only remaining problem: because of the above fix now compound sum and
compound subtraction catch the define polynomial function and return
wrong results. I think the solution would be changing the regex for
isPolynomial.